Start file system refactoring.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1416 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 8d961145b985ae8ead5d449dad12e3f180f8c6bd

+++ Source/ScriptExtensions.cpp
@@ -70,7 +70,7 @@ static uint CScriptArray_RemoveAll( CScriptArray* arr, void* value )
 {
     uint count = 0;
     int  index = 0;
-    while( index < arr->GetSize() )
+    while( index < (int) arr->GetSize() )
     {
         index = arr->Find( index, value );
         if( index != -1 )
